sistana: rename module {track => mix}
