+++ src/arclet/alconna/sistana/model/pattern.py
@@ -10,7 +10,7 @@ from elaina_triehard import TrieHard
 from arclet.alconna._dcls import safe_dcls_kw
 
 from .pointer import Pointer
-from .track import Preset
+from .mix import Preset
 from .snapshot import AnalyzeSnapshot, SubcommandTraverse
 
 if TYPE_CHECKING:
